Rework shadow-catcher darkening formula and expose controls as sliders
Replace the shadow catcher's darkness handling with a unified
visibility^exponent curve applied consistently to both the direct
illusion term and its continuation-bounce throughput, so the darkness
dial actually controls contact-shadow strength without crushing open
floor or losing effect at the extremes. The exponent itself is now
geometrically (not linearly) mapped from the darkness slider, giving a
log-style feel with more range near the top.

Also swap the Shadow Darkness/Metallic/Roughness spin boxes in the
Visualization/Environment panel for sliders with a live value label,
matching the panel's existing slider+label convention.

+++ include/VisualizationEnvironmentPanel.h
@@ -144,10 +144,10 @@ public slots:
 	void onGroundModeChanged();
 	void onFloorTextureStateChanged(bool checked);
 	void onReflectionsChanged(bool checked);
-	void onShadowDarknessChanged(double value);
+	void onShadowDarknessChanged(int value);
 	void onShadowCatcherColorClicked();
-	void onShadowCatcherMetallicChanged(double value);
-	void onShadowCatcherRoughnessChanged(double value);
+	void onShadowCatcherMetallicChanged(int value);
+	void onShadowCatcherRoughnessChanged(int value);
 	void onEnvMappingChanged(bool checked);
 	void onFloorOffsetChanged(double value);
 	void onRepeatSChanged(double value);
